Speed scales are now continuous

+++ gebr/gebr-job.c
@@ -34,7 +34,7 @@ struct _GebrJobPriv {
 	GebrCommJobStatus status;
 	gchar *run_type;
 
-	gint exec_speed;
+	gdouble exec_speed;
 	gchar *title;
 	gchar *queue;
 	gchar *hostname;
@@ -589,7 +589,7 @@ gebr_job_get_exec_speed(GebrJob *job)
 }
 
 void
-gebr_job_set_exec_speed(GebrJob *job, gint exec_speed)
+gebr_job_set_exec_speed(GebrJob *job, gdouble exec_speed)
 {
 	job->priv->exec_speed = exec_speed;
 }
